aiagent/searchagent/requirements.txt
langchain==0.1.0
langchain-core==0.1.8
langchain-openai==0.0.2
python-dotenv==1.0.0
duckduckgo-search==4.1.1
